requirements.txt.txt
pandas==1.5.0
yfinance==0.1.74
google-cloud-bigquery==3.3.3
google-cloud-storage==2.5.0